// MovieBooking/MovieBookingSPA/src/app/projection/domain/models/projection.model.ts
export interface IProjection{
    id: string;
    movieId: string;
    movieTitle: string;
    runtime: string;
    theaterHallId: string;
    theaterHallName: string;
    projectionDate: string;
    projectionTerm: string;
    numberOfSeats: number;
    numberOfReservedSeats: number;
    price: number;
}